test/mochitest/test_check_environment.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
  <div class="dialog error" id="errorDisplay">
    <section class="unknown">
      An unknown error has been reported. Please check WebConsole and JSConsole.
    </section>
    <section class="adbhelper">
      Please verify ADB Helper v0.8.6 is installed and enabled.
    </section>
    <section class="gecko-tooold">
      Your Firefox Nightly is too old, please update.
    </section>
    <section class="gecko-unsupported-os-arch">
      This addon is only targetting Firefix Nightly running on Linux 32/64 bits or OSX 64 bits.
    </section>
    <section class="gecko-unsupported-build">
      This addon is only targetting Firefox Nightly builds.
    </section>
  </div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  let errorElements = [ ".unknown", ".adbhelper", ".gecko-tooold",
                        ".gecko-unsupported-os-arch",
                        ".gecko-unsupported-build",
                        "#errorDisplay" ];

  function getElementComputedStyle(selector) {
    return document.defaultView.getComputedStyle(document.querySelector(selector));
  }

  function assert_no_error_displayed() {
    errorElements.forEach(msgClass => {
      let sect = getElementComputedStyle(msgClass);
      is(sect.display,    "none",   "error element " + msgClass + " is not displayed");
      is(sect.visibility, "hidden", "error element " + msgClass + " is not visible");
    });
  }

  function assert_unknown_error_displayed() {
    let ed = getElementComputedStyle("#errorDisplay");
    is(ed.display,    "flex",    "error screen is displayed");
    is(ed.visibility, "visible", "error screen is visible");

    let sect = getElementComputedStyle(".unknown");
    is(sect.display,    "block",   "unknown error is displayed");
    is(sect.visibility, "visible", "unknown error is visible");
  }

  function reset_error_displayed() {
    errorElements.forEach(msgClass => {
      document.querySelector(msgClass).style.display     = "none";
      document.querySelector(msgClass).style.visibility  = "hidden";
    });
  }

  // A promise rejection of verifyGecko() should display unknown error
  function test_verify_gecko() {
    is(typeof verifyGecko, "function", "has verifyGecko function");

    let mockVerifyGecko = function() {
      return new Promise((resolve, reject) => {
        return reject(new Error("GECKO_UNKNOWN"));
      });
    };

    verifyGecko = mockVerifyGecko;

    return new Promise((resolve, reject) => {
      return ensureEnvironment().then(() => resolve());
    });
  }

  // A promise rejection of verifyADBHelper() should display unknown error
  function test_verify_adbhelper() {
    is(typeof verifyADBHelper, "function", "has verifyADBHelper function");

    let mockVerifyADBHelper = function() {
      return new Promise((resolve, reject) => {
        return reject(new Error("ADBHELPER_UNKNOWN"));
      });
    };

    verifyADBHelper = mockVerifyADBHelper;

    return new Promise((resolve, reject) => {
      return ensureEnvironment().then(() => resolve());
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => reset_error_displayed())
      .then(() => assert_no_error_displayed())
      .then(() => reset_error_displayed())
      .then(() => test_verify_gecko())
      .then(() => assert_unknown_error_displayed())
      .then(() => reset_error_displayed())
      .then(() => test_verify_adbhelper())
      .then(() => assert_unknown_error_displayed())
      .then(() => reset_error_displayed())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
